requirements.txt
google-adk
requests
pandas
openpyxl
urllib3
